backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
fredapi==0.5.1
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
requests==2.31.0
